Add folder check to some features which were missing

+++ scripts/features/solr.sh
@@ -1,5 +1,11 @@
 #!/usr/bin/env bash
 
+if [ -f /home/vagrant/.homestead-features/solr ]
+then
+    echo "solr already installed."
+    exit 0
+fi
+
 # Install Java Runtime Enviroment
 sudo apt update
 sudo apt install default-jre php-solr -y
